Move ShadowyInsight from Features to Spells
