+++ src/analysis/retail/priest/shadow/modules/spells/ShadowyInsight.tsx
